Fix issue when JSON msg is sent and number of tokens is <1000
Issue was that `_sendJSONRequests()` always uses `async.forEachSeries()` to iterate over arrays of registration ids. So `registration_ids` parameter there should be `String[][]`.

However before this fix, a `String[]` would be passed. Then `async.forEachSeries()` would take this array and send message for each of contained `registration_ids`. And request would be invalid - Google expects to get an array of ids, not a String as in `registration_ids: "<reg id>"`.